requirements.dev.txt
PyYAML==3.12
gevent==1.2.1
transitions==0.4.3
nose==1.3.7